# backend/movieframe/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8443
  description: Generated server url
paths:
...
  /api/movies/{id}/review/{userName}:
    post:
      tags:
      - review-rest-controller
      summary: Post review into a movie
      operationId: newReviewAPI
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: userName
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Review'
        required: true
      responses:
        "404":
          description: No movie with this id was found to post a review
        "400":
          description: Invalid id supplied
        "200":
          description: Posted review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
...
  /reviews:
    get:
      tags:
      - review-rest-controller
      operationId: getReviews
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageReview'
  /reviews/{id}:
    get:
      tags:
      - review-rest-controller
      operationId: deleteReviewById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageReview'
  /reviews/user/{userName}:
    get:
      tags:
      - review-rest-controller
      operationId: getUserReviews
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PageReview'
...
  /api/reviews:
    get:
      tags:
      - review-rest-controller
      summary: Get reviews
      operationId: getReviewsAPI
      parameters:
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "200":
          description: Found reviews list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: No review founded
  /api/reviews/{id}:
    get:
      tags:
      - review-rest-controller
      summary: Get one review
      operationId: getReviewById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: No review with this id was found to delete
    delete:
      tags:
      - review-rest-controller
      summary: Delete review
      operationId: deleteReviewById_1
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Deleted review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        "404":
          description: No review with this id was found to delete
  /api/reviews/user/{userName}:
    get:
      tags:
      - review-rest-controller
      summary: Get the currently logged user reviews
      operationId: getUserReviewsAPI
      parameters:
      - name: userName
        in: path
        required: true
        schema:
          type: string
      - name: page
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/Pageable'
      responses:
        "400":
          description: Invalid username supplied
        "404":
          description: No review posted
        "200":
          description: Found user reviews
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
...
  /api/directors/{id}:
    get:
      tags:
      - director-rest-controller
      summary: Get director
      operationId: getDirector
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid id supplied
        "200":
          description: Found director
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Director'
        "404":
          description: No director with this id was found
components:
  schemas:
    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        rating:
          type: integer
          format: int32
        coments:
          type: string
        coment:
          type: string
          writeOnly: true
...
    Director:
      type: object
      properties:
        id:
          type: integer
          format: int64
        director:
          type: string
        biography:
          type: string
        name:
          type: string
        born:
          type: string
        residence:
          type: string
        score:
          type: number
          format: double
        imageFile:
          type: object
          properties:
            binaryStream:
              type: object
        genre:
          type: array
          items:
            type: string
...
    Pageable:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: '#/components/schemas/Sort'
        pageNumber:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        paged:
          type: boolean
        unpaged:
          type: boolean
    Sort:
      type: object
      properties:
        sorted:
          type: boolean
        unsorted:
          type: boolean
        empty:
          type: boolean
    PageReview:
      type: object
      properties:
        totalPages:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        size:
          type: integer
          format: int32
        content:
          type: array
          items:
            $ref: '#/components/schemas/Review'
        number:
          type: integer
          format: int32
        sort:
          $ref: '#/components/schemas/Sort'
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        pageable:
          $ref: '#/components/schemas/Pageable'
        empty:
          type: boolean
...